refactor: token entity 구조 변경 (#24)
