+++ src/main/java/project/backend/repository/auth/BlacklistTokenRedisRepository.java
@@ -1,7 +1,7 @@
 package project.backend.repository.auth;
 
 import org.springframework.data.repository.CrudRepository;
-import project.backend.security.token.BlacklistToken;
+import project.backend.entity.token.BlacklistToken;
 
 public interface BlacklistTokenRedisRepository extends CrudRepository<BlacklistToken, String> {
