+++ src/main/java/project/backend/repository/auth/RefreshTokenRedisRepository.java
@@ -2,7 +2,7 @@ package project.backend.repository.auth;
 
 import org.springframework.data.redis.core.RedisHash;
 import org.springframework.data.repository.CrudRepository;
-import project.backend.security.token.RefreshToken;
+import project.backend.entity.token.RefreshToken;
 
 @RedisHash
 public interface RefreshTokenRedisRepository extends CrudRepository<RefreshToken, Long> {
